backend/requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
ipykernel
mem0ai
pinecone
tavily-python
psycopg2-binary
pandas
openpyxl
sqlalchemy
fastapi
uvicorn
pydantic[email]
python-jose[cryptography]
bcrypt
bcrypt
pandas
python-multipart
langchain-text-splitters
mem0ai
langchain_pinecone
pinecone
langchain_community
unstructured[excel]
networkx
msoffcrypto-tool
reportlab